Implement DUMPQ/DUMPA, resolving #251.

+++ ccs_workspace/qc16_common/qc16.h
@@ -86,7 +86,7 @@ typedef struct {
     element_type element_selected;
     uint8_t element_level[3];
     uint8_t element_level_progress[3];
-    uint16_t element_qty[3];
+    uint32_t element_qty[3];
     uint16_t qbadges_connected_count;
     uint16_t cbadges_connected_count;
     uint16_t qbadge_max_id;
